refactor-full-config: rename
